refactor(flowmeters): Remove flowmeter list management, introduce FlowmeterWidgets
* Removed unused imports and state variables for flowmeters
* Replaced SensorStatistic flowmeter rendering with new FlowmeterWidgets component
* Updated useEffect hooks to only manage actuators listing

issue #29
